requirements.txt
pytest
pymongo
mysql-connector-python